pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maylang-cli"
version = "0.1.0"
description = "MayLang – Explainable Change Standard CLI for reducing AI tech debt."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "MayLang Contributors"},
]
keywords = ["maylang", "change-management", "explainability", "ai-tech-debt", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[project.scripts]
may = "maylang_cli.cli:main"

[project.urls]
Homepage = "https://github.com/maylang/maylang-cli"
Repository = "https://github.com/maylang/maylang-cli"
Issues = "https://github.com/maylang/maylang-cli/issues"

[tool.setuptools.packages.find]
include = ["maylang_cli*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
